+++ CHANGELOG.md
@@ -84,6 +84,8 @@
 - BZip3 (BZ3)
 - Broad Band eBook (BBeB)
 - Design Web Format (DWF)
+- FictionBook (FB2)
+- FictionBook Zipped (FBZ)
 - Flexible and Interoperable Data Transfer (FIT)
 - Game Gear ROM (GG)
 - IFF 8-Bit Sampled Voice (8SVX)